+++ src/lib.rs
@@ -98,16 +98,21 @@
 
 pub mod ast;
 mod combinator;
+#[cfg(feature = "expand-into-regex")]
+pub mod expand;
 pub mod parse;
 
 // TODO: Remove once `derive_more` 0.99.17 is released.
 use syn as _;
 
 #[doc(inline)]
-pub use self::{
-    ast::{
-        Alternation, Alternative, Expression, Optional, Parameter,
-        SingleAlternation, SingleExpression, Spanned,
-    },
-    parse::Error,
+pub use self::ast::{
+    Alternation, Alternative, Expression, Optional, Parameter,
+    SingleAlternation, SingleExpression, Spanned,
+};
+#[cfg(feature = "expand-into-regex")]
+#[doc(inline)]
+pub use self::expand::{
+    ExpansionError as Error, IntoRegexCharIter, ParametersProvider,
+    WithParameters,
 };